+++ src/loader/story.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
 
 import { storiesOf } from '@storybook/react'
-import { withKnobs, boolean, number } from '@storybook/addon-knobs'
+import { withKnobs, boolean, number, select } from '@storybook/addon-knobs'
 import Section from 'layout/section/baseSection'
-import Loader from './index'
+import Loader, { LoaderLayoutMode } from './index'
 
 const stories = storiesOf('Widgets|Loader', module)
 stories.addDecorator(withKnobs)
@@ -11,7 +11,18 @@ stories.addDecorator(withKnobs)
 stories.add('default', () => (
   <Section>
     <Loader
-      inline={boolean('inline', true)}
+      layoutMode={select('layoutMode', LoaderLayoutMode, LoaderLayoutMode.FULLSCREEN)}
+      size={number('size', 48)}
+      done={boolean('done', false)}
+    />
+  </Section>
+))
+
+stories.add('legacy inline prop', () => (
+  <Section>
+    <Loader
+      inline={boolean('inline (deprecated - use layoutMode)', true)}
+      layoutMode={select('layoutMode', LoaderLayoutMode, null)}
       size={number('size', 48)}
       done={boolean('done', false)}
     />